moving syncChanedFiles to common location
