+++ lua/grug-far/engine/ripgrep.lua
@@ -2,7 +2,7 @@ local fetchCommandOutput = require('grug-far/engine/fetchCommandOutput')
 local parseResults = require('grug-far/engine/ripgrep/parseResults')
 local fetchFilesWithMatches = require('grug-far/engine/ripgrep/fetchFilesWithMatches')
 local replaceInMatchedFiles = require('grug-far/engine/ripgrep/replaceInMatchedFiles')
-local syncChangedFiles = require('grug-far/engine/ripgrep/syncChangedFiles')
+local syncChangedFiles = require('grug-far/engine/syncChangedFiles')
 local getArgs = require('grug-far/engine/ripgrep/getArgs')
 local colors = require('grug-far/engine/ripgrep/colors')
 local utils = require('grug-far/utils')
